+++ extendr-api/tests/optional/either/either_into_robj.rs
