+++ extendr-api/tests/optional/either/either_try_from_robj.rs
