+++ extendr-api/tests/optional/either/mod.rs
@@ -1,2 +1,2 @@
-mod either_robj_from_tests;
-mod either_try_from_tests;
+mod either_into_robj;
+mod either_try_from_robj;
